Change kick strength type

+++ behavior-ms/behavior-bruxo/behavior/processing/messages/common/robot_kick/kick_command.cpp
@@ -2,7 +2,7 @@
 
 namespace behavior {
 
-KickCommandMessage::KickCommandMessage(double strength,
+KickCommandMessage::KickCommandMessage(float strength,
                                        bool is_front,
                                        bool is_chip,
                                        bool charge_capacitor,
